added installment logic

+++ divinespark-backend/src/main/java/com/divinespark/dto/DonationCreateRequest.java
@@ -1,11 +1,11 @@
 package com.divinespark.dto;
 
 public class DonationCreateRequest {
-    private double amount;
+    private int amount;
     private String note;
 
-    public double getAmount() { return amount; }
-    public void setAmount(double amount) { this.amount = amount; }
+    public int getAmount() { return amount; }
+    public void setAmount(int amount) { this.amount = amount; }
     public String getNote() { return note; }
     public void setNote(String note) { this.note = note; }
 }
