+++ divinespark-backend/src/main/java/com/divinespark/service/RazorpayService.java
@@ -3,5 +3,5 @@ package com.divinespark.service;
 import com.divinespark.dto.RazorpayOrderResponse;
 
 public interface RazorpayService {
-    RazorpayOrderResponse createOrder(double amount, Long bookingId);
+    RazorpayOrderResponse createOrder(int amount, Long bookingId);
 }
